requirements.txt
torch>=2.5.1
torchvision>=0.20.1
torchaudio>=2.5.1
transformers>=4.40.0
peft>=0.10.0
datasets
pyyaml
accelerate
safetensors
scipy
bitsandbytes>=0.43.0
